pentaho_reports.
Add method 'pentaho_report_action' to the ir.actions.report.xml model.
This method returns the action definition to run a Pentaho report.

This method can be used from ad-hoc wizards or server actions to trigger
Pentaho reports.

The action definition is returned as a dict which can be returned
to the OpenERP client from a wizard button or server action to
cause the client to request the report.