Changes due to cache flushing.
Creating users / partners on the fly was causing a problem in on change on email template.

Various interactions between old api and new api within standard odoo user creations had code which
invalidated the cache.  This was a problem for the wizard cache which called the onchange which
called the report in preview mode.

Some of the fix is nasty code, BUT it is autonomous...

We are still looking for a better and cleaner solution, long term.